Add files via upload
In this project basically we have analyse the weather forcast in the Delhi region by the help of Dataset. We have used pandas,numpy,plotly,matplotlib. In this project we have learned a new Model which is known as Prophet. This model helps us to predict the time-series problems. This model  is of Facebook they also use it to predict the weather foracst for couple of days